requirements.txt
pyautogui
pytesseract
Pillow
googletrans==4.0.0-rc1
overlay
pygame
keyboard

